refactor: extract direct uploaded context guards
